AOO-27: passcode feature - new files for hashing
