+++ app/src/main/java/net/opendasharchive/openarchive/features/settings/passcode/HashingStrategy.kt
@@ -1,0 +1,7 @@
+package net.opendasharchive.openarchive.features.settings.passcode
+
+interface HashingStrategy {
+    suspend fun hash(passcode: String, salt: ByteArray): ByteArray
+    suspend fun generateSalt(): ByteArray
+    val saltLength: Int
+}
